+++ flowtype/commands/__init__.py
@@ -1,7 +1,9 @@
 from .add_pragma import FlowtypeAddPragma
 from .check_contents import FlowtypeCheckContents
+from .goto_definition import FlowtypeGotoDefinition
 
 __all__ = [
     'FlowtypeAddPragma',
     'FlowtypeCheckContents',
+    'FlowtypeGotoDefinition',
 ]